+++ lib/sources/foursquare.rb
@@ -1,14 +1,14 @@
 # frozen_string_literal: true
 
-require_relative '../parsers/ics'
+require_relative '../parsers/kml'
 
 class Foursquare
   def initialize(root:, type:)
-    filename = ::Dir.glob("#{root}/*.ics").
+    filename = ::Dir.glob("#{root}/*.kml").
       last
 
-    @data = ::Parser::ICS.new(path: filename).
-      geojson
+    @data = ::Parser::KML.new(path: filename, type: :point).
+      data
   end
 
   def geojson